Prevent stack overflow when evaluating input arguments
